fix: BUG-7599 - Issue 2, 4
