+++ src/components/helpers/utils.ts
@@ -1,3 +1,5 @@
+import { getSdkConfig } from '@pega/react-sdk-components/lib/components/helpers/config_access';
+
 export const scrollToTop = () => {
   const position = document.getElementById('#main-content')?.offsetTop || 0;
   document.body.scrollTop = position;
@@ -14,4 +16,33 @@ export const checkErrorMsgs = (errorMsgs = [], fieldIdentity = '', fieldElement 
     element =>
       element.message.fieldId === fieldIdentity || element.message.fieldId.startsWith(fieldElement)
   );
+};
+
+export const getServiceShutteredStatus = async (): Promise<boolean> => {
+  interface ResponseType {
+    data: { Shuttered: boolean };
+  }
+  try {
+    const sdkConfig = await getSdkConfig();
+    const urlConfig = new URL(
+      `${sdkConfig.serverConfig.infinityRestServerUrl}/app/${sdkConfig.serverConfig.appAlias}/api/application/v2/data_views/D_ShutterLookup`
+    ).href;
+    const featureID = 'ChB';
+    const featureType = 'Service';
+
+    const parameters = new URLSearchParams(
+      `{FeatureID: ${featureID}, FeatureType: ${featureType}}`
+    );
+
+    const url = `${urlConfig}?dataViewParameters=${parameters}`;
+    const { invokeCustomRestApi } = PCore.getRestClient();
+    const response: ResponseType = await invokeCustomRestApi(url, {
+      method: 'GET'
+    });
+    return response.data.Shuttered;
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.log(error);
+    return false;
+  }
 };
